[#56][be][notification]refactor: 알림 로그로 도메인명 변경
