+++ back/src/main/java/com/rouby/notification/application/facade/notificationlog/NotificationLogFacade.java
@@ -1,10 +1,10 @@
-package com.rouby.notification.application.facade;
+package com.rouby.notification.application.facade.notificationlog;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class NotificationFacade {
+public class NotificationLogFacade {
 
 }
