+++ back/src/main/java/com/rouby/notification/domain/repository/notificationlog/NotificationLogRepository.java
@@ -1,0 +1,5 @@
+package com.rouby.notification.domain.repository.notificationlog;
+
+public interface NotificationLogRepository {
+
+}
