+++ back/src/main/java/com/rouby/notification/infrastructure/persistence/jpa/notificationlog/NotificationLogJpaRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.rouby.notification.infrastructure.persistence.jpa.notificationlog;
+
+public interface NotificationLogJpaRepositoryCustom {
+
+}
