+++ back/src/main/java/com/rouby/notification/infrastructure/persistence/jpa/notificationlog/NotificationLogJpaRepositoryCustomImpl.java
@@ -1,0 +1,11 @@
+package com.rouby.notification.infrastructure.persistence.jpa.notificationlog;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class NotificationLogJpaRepositoryCustomImpl implements NotificationLogJpaRepositoryCustom {
+
+  private final JPAQueryFactory jpaQueryFactory;
+
+}
